Refactor classes into python module namespaces
